provider/handlers/keystone.py
# Copyright 2016 Red Hat, Inc.
#
# This program is free software; you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation; either version 2 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program; if not, write to the Free Software
# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
#
# Refer to the README and COPYING files for full details of the license
#
from __future__ import absolute_import

import json

from handlers.selecting_handler import SelectingHandler
from handlers.keystone_responses import responses


class TokenHandler(SelectingHandler):

    def call_response_handler(self, response_handler, content, id):
        return json.dumps(response_handler(
            json.loads(content) if content else {}, id))

    @staticmethod
    def get_responses():
        return responses()
